Update Anton and Digits.cpp

+++ 800/Anton and Digits.cpp
@@ -5,7 +5,7 @@ using namespace std;
 	
 	int main()
 	{
-		int a,b,c,d;	
+		ll a,b,c,d;	
 		cin >> a >> b >> c >> d;
 		int kp = min(a, min(c, d));
 		cout << kp * 256 + min(a - kp, b) * 32;
